requirements.txt
Flask-PyMongo
Flask-Session
python-dotenv
Flask-Mail
pymongo
dnspython
azure-storage-blob 
azure-identity